schema: refactor taxonomy-term
