+++ __fixtures__/uuid.ts
@@ -19,6 +19,8 @@
  * @license   http://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://github.com/serlo-org/api.serlo.org for the canonical source repository
  */
+import { gql } from 'apollo-server'
+
 import { Instance } from '../src/graphql/schema/instance'
 import {
   AliasPayload,
@@ -50,6 +52,39 @@ import {
 } from '../src/graphql/schema/uuid'
 import { NavigationPayload } from '../src/graphql/schema/uuid/navigation'
 import { license } from './license'
+
+export function createSetTaxonomyTermMutation(variables: TaxonomyTermPayload) {
+  return {
+    mutation: gql`
+      mutation setTaxonomyTerm(
+        $id: Int!
+        $trashed: Boolean!
+        $alias: String
+        $type: TaxonomyTermType!
+        $instance: Instance!
+        $name: String!
+        $description: String
+        $weight: Int!
+        $parentId: Int
+        $childrenIds: [Int!]!
+      ) {
+        _setTaxonomyTerm(
+          id: $id
+          trashed: $trashed
+          alias: $alias
+          type: $type
+          instance: $instance
+          name: $name
+          description: $description
+          weight: $weight
+          parentId: $parentId
+          childrenIds: $childrenIds
+        )
+      }
+    `,
+    variables,
+  }
+}
 
 export const navigation: NavigationPayload = {
   data: JSON.stringify([
